Move fetchLatestJson to updateApp
