+++ app/renderer/src/AppCycle/startUp.ts
@@ -2,7 +2,7 @@ import SETTINGS from '../../store/Settings'
 import UpdateWindow from '../../windows/NewUpdateWindow'
 import DiscordWindow from '../../windows/Discord'
 import { shell } from 'electron'
-import fetchLatestJson from '../../src/Utilities/FetchLatestJson'
+import { fetchLatestJson } from '../../src/Utilities/updateApp'
 import { CONNECTION } from '../../src/Utilities/ConnectionStatus'
 import { setupDefaultMenus } from '../UI/Toolbar/setupDefaults'
 import { createNotification } from '../UI/Footer/create'
